refactor: move components closer to usage
