+++ src/routes/TradingLeague/Points/Points.tsx
