+++ src/routes/TradingLeague/RankingHeader/RankingHeader.tsx
